change id type in wlocation domain

+++ domain/src/main/java/com/mistpaag/domain/WLocation.kt
@@ -3,5 +3,5 @@ package com.mistpaag.domain
 data class WLocation(
     val title: String,
     val locationType: String,
-    val woeid: Int
+    val woeid: Long
 )
